requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# gesture-controller-firmware/firmware.py: 8
Pillow == 5.4.1

# gesture-controller-firmware/firmware.py: 11
# gesture-controller-firmware/screen_device.py: 2
luma.core == 2.3.1

# gesture-controller-firmware/firmware.py: 14
# gesture-controller-firmware/test.py: 4
mido == 1.2.10

# gesture-controller-firmware/firmware.py: 4
# gesture-controller-firmware/test.py: 2
skywriter == 0.0.7
